requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
deepface==0.0.79
distlib==0.3.7
Django==4.2.1
djangorestframework==3.14.0
filelock==3.13.1
fire==0.5.0
Flask==2.3.2
flatbuffers==23.3.3
gast==0.4.0
gdown==4.7.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
keras-vggface @ git+https://github.com/rcmalli/keras-vggface.git@bee35376e76e35d00aeec503f2f242611a97b38a
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
mtcnn==0.1.1
mysql-connector-python-rf==2.2.2
nudenet==3.0.8
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.16.1
opencv-python==4.7.0.72
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pipenv==2023.10.24
platformdirs==3.11.0
protobuf==3.20.*
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.24.6
Werkzeug==2.3.3
wrapt==1.14.1
zipp==3.15.0
requests==2.31.0
psycopg2-binary==2.9.9
django-storages==1.12.3
google-api-core==2.7.2
google-cloud-core==2.3.0
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0